requirements.txt
aiogram==3.7.0
httpx==0.27.0
phonenumbers==8.13.39
python-whois==0.9.4
python-dotenv==1.0.1
sherlock-project==0.14.4
maigret==0.4.4
holehe==1.4.4
